fix(runtime): preserve direct image execution
